# src/swagger.yaml
openapi: 3.0.1
info:
  title: Seed API microservice
  version: 1.0.0
servers:
  - url: /api/project/service/v1
security:
  - ApiKeyAuth: []
paths:
  /grafana:
    x-swagger-router-controller: grafana
    get:
      x-anonymous: true
      tags:
        - Grafana
      description: Grafana json data source health probe
      operationId: getHealth
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OptionResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    OptionResponse:
      description: Equity Derivative options
      type: object
...
    ErrorResponse:
      success:
        type: boolean
      message:
        type: string
  securitySchemes:
    ApiKeyAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header # can be "header", "query" or "cookie"
      name: x-api-key